#96 : Sam Cusworth : Update socrates to adhere to LFRic compiler checks : task : general : regression : minor Merged into /main/trunk: /main/branches/dev/samcusworth/um11.9_lfric_checks@994 cf. /main/trunk@989
git-svn-id: https://metomi/svn/socrates.xm/main/trunk@995 7a691fcb-5d1e-4fc1-8b6b-54125fa6be09

+++ src/radiance_core/circumsolar_fraction.F90
@@ -66,7 +66,7 @@ SUBROUTINE circumsolar_fraction(n_profile                               &
 !
 ! Calculate index of Legedre polynomial based corresponding to
 ! the input half viewing angle of pyrheliometer 
-  ind_ang=half_angle/0.25
+  ind_ang=INT(half_angle/0.25)
   one_minus_cos = 1.0 - cos_half_angle(ind_ang)
   p_legendre_2d = reshape(p_legendre, (/32,20/))
   forward_scatter_csr=0.0
